image-to-ascii/api/requirements.txt
Flask==2.0.1
Pillow==8.3.1
numpy==1.21.1
flask-cors==3.0.10